SYSREF request moved from AD9528 to application
The AD9528 registers are configured via SPI in such a way that
the SYSREF is done via GPIO between FPGA and AD9528.

The GPIO control is therefore application specific and not
AD9528 driver specific.
