+++ projects/adrv9009/src/app/headless.c
@@ -45,6 +45,7 @@
 #include "talise_arm_binary.h"
 #include "talise_stream_binary.h"
 #include "app_config.h"
+#include "app_sysref_req.h"
 
 /**********************************************************/
 /**********************************************************/
@@ -55,8 +56,9 @@
 int main(void)
 {
 	adiHalErr_t ad9528Error;
+	struct gpio_desc *gpio_sysref_req;
+
 	ad9528Device_t clockAD9528_ = {
-		NULL,
 		NULL,
 		NULL,
 		&clockSpiSettings,
@@ -290,6 +292,11 @@ int main(void)
 	 * to the Talise device.
 	 **/
 
+	if(SUCCESS != gpio_get(&gpio_sysref_req, ADRV_SYSREF_REQ)) {
+		printf("error: gpio_get() failed for pin %d\n", ADRV_SYSREF_REQ);
+		goto error_0;
+	}
+
 	/* Init the AD9528 data structure */
 	ad9528Error = AD9528_initDeviceDataStruct(clockAD9528_device,
 			clockAD9528_device->pll1Settings->vcxo_Frequency_Hz,
@@ -521,7 +528,7 @@ int main(void)
 	}
 
 	/*< user code - Request minimum 3 SYSREF pulses from Clock Device - > */
-	AD9528_sysrefReq(clockAD9528_device, SYSREF_PULSE);
+	sysrefReq(gpio_sysref_req, SYSREF_PULSE);
 
 	/*******************/
 	/**** Verify MCS ***/
@@ -691,7 +698,7 @@ int main(void)
 
 	/*** < User Sends SYSREF Here > ***/
 
-	AD9528_sysrefReq(clockAD9528_device, SYSREF_CONT_ON);
+	sysrefReq(gpio_sysref_req, SYSREF_CONT_ON);
 
 	mdelay(100);
 
@@ -700,7 +707,7 @@ int main(void)
 
 	mdelay(100);
 
-	AD9528_sysrefReq(clockAD9528_device, SYSREF_CONT_OFF);
+	sysrefReq(gpio_sysref_req, SYSREF_CONT_OFF);
 
 	mdelay(100);
 
@@ -765,7 +772,7 @@ int main(void)
 		goto error_10;
 	}
 
-	AD9528_sysrefReq(clockAD9528_device, SYSREF_CONT_ON);
+	sysrefReq(gpio_sysref_req, SYSREF_CONT_ON);
 
 	axi_jesd204_rx_watchdog(rx_jesd);
 	axi_jesd204_rx_watchdog(rx_os_jesd);
@@ -844,6 +851,8 @@ int main(void)
 	axi_clkgen_remove(rx_clkgen);
 #endif
 	AD9528_remove(clockAD9528_device);
+	gpio_remove(gpio_sysref_req);
+
 
 	printf("Bye\n");
 
@@ -883,6 +892,7 @@ error_2:
 #endif
 error_1:
 	AD9528_remove(clockAD9528_device);
+	gpio_remove(gpio_sysref_req);
 error_0:
 	return FAILURE;
 }
